requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.6
python-dotenv==1.0.0
docker==6.1.3 
=======
python-telegram-bot==21.4
docker==7.1.0
python-dotenv==0.19.0
>>>>>>> 068b126cfa7522ee24bf222eda4cf09fdbc90332
